Add elasticsearch implementation
